+++ server/db/db.js
@@ -1,0 +1,6 @@
+const Sequelize = require('sequelize')
+const pkg = require('../../package.json')
+
+const db = new Sequelize(process.env.DATABASE_URL || `postgres://localhost/${pkg.name}`, {logging: false})
+
+module.exports = db
