+++ server/db/models/index.js
@@ -1,0 +1,7 @@
+const Message = require('./message')
+const User = require('./users')
+
+Message.belongsTo(User)
+User.hasMany(Message)
+
+module.exports = {Message, User}
