+++ server/db/models/message.js
@@ -1,0 +1,10 @@
+const Sequelize = require('sequelize')
+const db = require('../db')
+
+const Message = db.define('message', {
+    body: {
+        type: Sequelize.TEXT,
+    }
+})
+
+module.exports = Message
